refactor(nlu): added middleware for legacy intent election
