+++ modules/ndu/src/backend/middleware.ts
@@ -6,7 +6,7 @@ export const registerMiddleware = async (bp: typeof sdk, bots: MountedBots) => {
   bp.events.registerMiddleware({
     name: 'ndu.incoming',
     direction: 'incoming',
-    order: 13,
+    order: 14,
     description: 'Where magic happens',
     handler: async (event: sdk.IO.IncomingEvent, next: sdk.IO.MiddlewareNextCallback) => {
       if (bots[event.botId]) {
