+++ modules/nlu-extras/src/backend/middleware.ts
@@ -18,7 +18,7 @@ export const registerMiddleware = async (bp: typeof sdk) => {
   bp.events.registerMiddleware({
     name: 'nlu-extras.incoming',
     direction: 'incoming',
-    order: 11,
+    order: 12,
     description:
       'Process natural language in the form of text. Structured data with an action and parameters for that action is injected in the incoming message event.',
     handler: async (event: sdk.IO.IncomingEvent, next: sdk.IO.MiddlewareNextCallback) => {
