+++ modules/nlu/src/backend/predict-pipeline.ts
@@ -7,7 +7,6 @@ import LanguageIdentifierProvider, { NA_LANG } from './language/language-identif
 import { isPOSAvailable } from './language/pos-tagger'
 import { getUtteranceFeatures } from './out-of-scope-featurizer'
 import SlotTagger from './slots/slot-tagger'
-import * as math from './tools/math'
 import { replaceConsecutiveSpaces } from './tools/strings'
 import { EXACT_MATCH_STR_OPTIONS, ExactMatchIndex, TrainArtefacts } from './training-pipeline'
 import { Intent, PatternEntity, SlotExtractionResult, Tools } from './typings'
@@ -59,8 +58,6 @@ type E1IntentPred = {
 
 const DEFAULT_CTX = 'global'
 const NONE_INTENT = 'none'
-const OOS_AS_NONE_TRESH = 0.3
-const LOW_INTENT_CONFIDENCE_TRESH = 0.4
 
 async function DetectLanguage(
   input: PredictInput,
@@ -280,132 +277,6 @@ async function predictIntent(input: PredictStep, predictors: Predictors): Promis
   }
 }
 
-// taken from svm classifier #295
-// this means that the 3 best predictions are really close, do not change magic numbers
-function predictionsReallyConfused(predictions: sdk.MLToolkit.SVM.Prediction[]): boolean {
-  if (predictions.length <= 2) {
-    return false
-  }
-
-  const std = math.std(predictions.map(p => p.confidence))
-  const diff = (predictions[0].confidence - predictions[1].confidence) / std
-  if (diff >= 2.5) {
-    return false
-  }
-
-  const bestOf3STD = math.std(predictions.slice(0, 3).map(p => p.confidence))
-  return bestOf3STD <= 0.03
-}
-
-function electIntent(input: PredictOutput): PredictOutput {
-  const allCtx = Object.keys(input.predictions)
-
-  const ctx_predictions = allCtx.map(label => {
-    const { confidence } = input.predictions[label]
-    return { label, confidence }
-  })
-
-  const oss_label = 'oos'
-  const oos_predictions = {
-    label: oss_label,
-    confidence: input.predictions[oss_label].confidence
-  }
-
-  const perCtxIntentPrediction = _.chain(input.predictions)
-    .mapValues(v => v.intents)
-    .value()
-
-  const totalConfidence = Math.min(
-    1,
-    _.sumBy(
-      ctx_predictions.filter(x => input.includedContexts.includes(x.label)),
-      'confidence'
-    )
-  )
-  const ctxPreds = ctx_predictions.map(x => ({ ...x, confidence: x.confidence / totalConfidence }))
-
-  // taken from svm classifier #349
-  let predictions = _.chain(ctxPreds)
-    .flatMap(({ label: ctx, confidence: ctxConf }) => {
-      const intentPreds = _.chain(perCtxIntentPrediction[ctx] || [])
-        .thru(preds => {
-          if (oos_predictions?.confidence > OOS_AS_NONE_TRESH) {
-            return [
-              ...preds,
-              {
-                label: NONE_INTENT,
-                confidence: oos_predictions?.confidence ?? 1,
-                context: ctx,
-                l0Confidence: ctxConf
-              }
-            ]
-          } else {
-            return preds
-          }
-        })
-        .map(p => ({ ...p, confidence: _.round(p.confidence, 2) }))
-        .orderBy('confidence', 'desc')
-        .value()
-
-      if (intentPreds[0].confidence === 1 || intentPreds.length === 1) {
-        return [{ label: intentPreds[0].label, l0Confidence: ctxConf, context: ctx, confidence: 1 }]
-      }
-
-      if (predictionsReallyConfused(intentPreds)) {
-        intentPreds.unshift({ label: NONE_INTENT, context: ctx, confidence: 1, slots: {} })
-      }
-
-      const lnstd = math.std(intentPreds.filter(x => x.confidence !== 0).map(x => Math.log(x.confidence))) // because we want a lognormal distribution
-      let p1Conf = math.GetZPercent((Math.log(intentPreds[0].confidence) - Math.log(intentPreds[1].confidence)) / lnstd)
-      if (isNaN(p1Conf)) {
-        p1Conf = 0.5
-      }
-
-      return [
-        {
-          label: intentPreds[0].label,
-          l0Confidence: ctxConf,
-          context: ctx,
-          confidence: _.round(ctxConf * p1Conf, 3)
-        },
-        {
-          label: intentPreds[1].label,
-          l0Confidence: ctxConf,
-          context: ctx,
-          confidence: _.round(ctxConf * (1 - p1Conf), 3)
-        }
-      ]
-    })
-    .orderBy('confidence', 'desc')
-    .filter(p => input.includedContexts.includes(p.context))
-    .uniqBy(p => p.label)
-    .map(p => ({ name: p.label, context: p.context, confidence: p.confidence }))
-    .value()
-
-  const ctx = _.get(predictions, '0.context', 'global')
-  const shouldConsiderOOS =
-    predictions.length &&
-    predictions[0].name !== NONE_INTENT &&
-    predictions[0].confidence < LOW_INTENT_CONFIDENCE_TRESH &&
-    oos_predictions?.confidence > OOS_AS_NONE_TRESH
-  if (!predictions.length || shouldConsiderOOS) {
-    predictions = _.orderBy(
-      [
-        ...predictions.filter(p => p.name !== NONE_INTENT),
-        { name: NONE_INTENT, context: ctx, confidence: oos_predictions?.confidence ?? 1 }
-      ],
-      'confidence'
-    )
-  }
-
-  const elected = _.maxBy(predictions, 'confidence')
-  return {
-    ...input,
-    intents: predictions,
-    intent: elected
-  }
-}
-
 async function predictOutOfScope(input: PredictStep, predictors: Predictors, tools: Tools): Promise<PredictStep> {
   if (!isPOSAvailable(input.languageCode) || !predictors.oos_classifier) {
     return input
@@ -425,22 +296,6 @@ async function predictOutOfScope(input: PredictStep, predictors: Predictors, too
   }
 }
 
-function detectAmbiguity(input: PredictOutput): PredictOutput {
-  // +- 10% away from perfect median leads to ambiguity
-  const preds = input.intents
-  const perfectConfusion = 1 / preds.length
-  const low = perfectConfusion - 0.1
-  const up = perfectConfusion + 0.1
-  const confidenceVec = preds.map(p => p.confidence)
-
-  const ambiguous =
-    preds.length > 1 &&
-    (math.allInRange(confidenceVec, low, up) ||
-      (preds[0].name === NONE_INTENT && math.allInRange(confidenceVec.slice(1), low, up)))
-
-  return { ...input, ambiguous }
-}
-
 async function extractSlots(input: PredictStep, predictors: Predictors): Promise<PredictStep> {
   const slots_per_intent: typeof input.slot_predictions_per_intent = {}
   for (const intent of predictors.intents.filter(x => x.slot_definitions.length > 0)) {
@@ -449,16 +304,6 @@ async function extractSlots(input: PredictStep, predictors: Predictors): Promise
   }
 
   return { ...input, slot_predictions_per_intent: slots_per_intent }
-}
-
-async function extractElectedIntentSlots(input: PredictOutput): Promise<PredictOutput> {
-  const intentWasElectedWithoutAmbiguity = input?.intent?.name && !_.isEmpty(input.predictions) && !input.ambiguous
-  if (!intentWasElectedWithoutAmbiguity) {
-    return input
-  }
-
-  const electedIntent = _.flatMap(input.predictions, p => p.intents).find(i => i.label === input.intent.name) // might find an intent with the correct name of an incorrect context... Chosen context should maybe be in PredictOuput
-  return { ...input, slots: electedIntent.slots } // no more range validation (in  utterance.tagSlot())
 }
 
 function MapStepToOutput(step: PredictStep, startTime: number): PredictOutput {
@@ -525,7 +370,6 @@ function MapStepToOutput(step: PredictStep, startTime: number): PredictOutput {
   )
 
   return {
-    ambiguous: step.intent_predictions.ambiguous,
     detectedLanguage: step.detectedLanguage,
     entities,
     errored: false,
@@ -579,11 +423,7 @@ export const Predict = async (
 
     stepOutput = await extractSlots(stepOutput, predictors)
 
-    let predictionOutput = MapStepToOutput(stepOutput, t0)
-    predictionOutput = electIntent(predictionOutput)
-    predictionOutput = detectAmbiguity(predictionOutput)
-    predictionOutput = await extractElectedIntentSlots(predictionOutput)
-    return predictionOutput
+    return MapStepToOutput(stepOutput, t0)
   } catch (err) {
     if (err instanceof InvalidLanguagePredictorError) {
       throw err
