+++ src/bp/sdk/botpress.d.ts
@@ -668,7 +668,7 @@ declare module 'botpress/sdk' {
     export interface EventUnderstanding {
       intent?: NLU.Intent
       /** Predicted intents needs disambiguation */
-      readonly ambiguous: boolean
+      readonly ambiguous?: boolean
       intents?: NLU.Intent[]
       /** The language used for prediction. Will be equal to detected language when its part of supported languages, falls back to default language otherwise */
       readonly language: string
